+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.7.9 - Unreleased
 
+### Fixed
+
+* Added workaround for an issue where a postgresql cache of a coco dataset is unable to remember its original json filepath.
+
 
 ## Version 0.7.8 - Released 2024-03-20